update authenticated route to /home to avoide routing complexity with landing page
