+++ app/components/AuthenticatedAndOnboardedRoute.tsx
@@ -19,7 +19,7 @@ export default function AuthenticatedAndOnboardedRoute({
   }, [session, router]);
 
   if (!session || !session.user) {
-    return null;
+    router.push('/login');
   }
 
   return <>{children}</>;
